Extract IngestService
